code review (refactor)
