+++ tests/common/regression/test_el_rewards.py
@@ -7,9 +7,8 @@ from functools import partial
 from brownie import interface, reverts, web3
 
 from utils.config import contracts, lido_dao_execution_layer_rewards_vault
-from utils.import_current_vote import get_start_and_execute_votes_func
+from utils.import_current_votes import is_there_any_vote_scripts, start_and_execute_votes
 
-start_and_execute_votes = get_start_and_execute_votes_func()
 
 LIDO_EXECUTION_LAYER_REWARDS_VAULT = lido_dao_execution_layer_rewards_vault
 TOTAL_BASIS_POINTS = 10000
@@ -36,8 +35,8 @@ def lido_execution_layer_rewards_vault():
     return interface.LidoExecutionLayerRewardsVault(LIDO_EXECUTION_LAYER_REWARDS_VAULT)
 
 
-@pytest.fixture(scope="module", autouse=(start_and_execute_votes is not None))
-def autoexecute_vote(vote_id_from_env, ldo_holder, helpers, accounts, dao_voting):
+@pytest.fixture(scope="module", autouse=is_there_any_vote_scripts())
+def autoexecute_vote(vote_id_from_env, helpers, accounts, dao_voting):
     if vote_id_from_env:
         helpers.execute_vote(
             vote_id=vote_id_from_env, accounts=accounts, dao_voting=dao_voting, topup='0.5 ether'
