+++ tests/common/regression/test_keys_op_index.py
@@ -8,9 +8,7 @@ import os
 import glob
 
 from brownie import web3
-from utils.import_current_vote import get_start_and_execute_votes_func
-
-start_and_execute_votes = get_start_and_execute_votes_func()
+from utils.import_current_votes import is_there_any_vote_scripts, start_and_execute_votes
 
 
 @pytest.fixture(scope="module")
@@ -18,7 +16,7 @@ def voting(accounts, dao_voting):
     return accounts.at(dao_voting.address, force=True)
 
 
-@pytest.fixture(scope="module", autouse=(start_and_execute_votes is not None))
+@pytest.fixture(scope="module", autouse=is_there_any_vote_scripts())
 def autoexecute_vote(vote_id_from_env, helpers, accounts, dao_voting):
     if vote_id_from_env:
         helpers.execute_vote(
