+++ tests/common/regression/test_staking_limits.py
@@ -7,10 +7,7 @@ import eth_abi
 from web3 import Web3
 from utils.config import contracts
 from brownie import web3, convert, reverts, ZERO_ADDRESS, chain
-from utils.import_current_vote import get_start_and_execute_votes_func
-
-
-start_and_execute_votes = get_start_and_execute_votes_func()
+from utils.import_current_votes import is_there_any_vote_scripts, start_and_execute_votes
 
 
 ether = 10 ** 18
@@ -26,7 +23,7 @@ def operator(accounts, dao_voting):
     return accounts.at(dao_voting.address, force=True)
 
 
-@pytest.fixture(scope="module", autouse=(start_and_execute_votes is not None))
+@pytest.fixture(scope="module", autouse=is_there_any_vote_scripts())
 def autoexecute_vote(vote_id_from_env, helpers, accounts, dao_voting):
     if vote_id_from_env:
         helpers.execute_vote(
